refactor: User api와 관련된 메서드들을 하나의 컨트롤러에서 관리하도록 수정
