+++ slipp/src/main/java/slipp/controller/exception/UserCreationFailException.java
@@ -1,0 +1,7 @@
+package slipp.controller.exception;
+
+public class UserCreationFailException extends RuntimeException {
+    public UserCreationFailException(Throwable cause) {
+        super(cause);
+    }
+}
